+++ backend/src/main/resources/db/migration/V1__init.sql
@@ -2,7 +2,7 @@ CREATE SEQUENCE hibernate_sequence;
 
 CREATE TABLE users
 (
-    id         int8         NOT NULL,
+    id         bigint       NOT NULL,
     email      varchar(255) NOT NULL,
     "password" varchar(255) NOT NULL,
     username   varchar(255) NOT NULL,
@@ -13,10 +13,10 @@ CREATE TABLE users
 
 CREATE TABLE baskets
 (
-    id           int8         NOT NULL,
+    id           bigint       NOT NULL,
     name         varchar(255) NOT NULL,
     description  varchar(255) NOT NULL,
-    owner_id     int8         NOT NULL,
+    owner_id     bigint       NOT NULL,
     created_date timestamp    NOT NULL,
     CONSTRAINT baskets_users_fk FOREIGN KEY (owner_id) REFERENCES users (id),
     CONSTRAINT baskets_pk PRIMARY KEY (id)
@@ -24,17 +24,17 @@ CREATE TABLE baskets
 
 CREATE TABLE investments
 (
-    id                   int8         NOT NULL,
-    owner_id             int8         NOT NULL,
-    basket_id            int8         NOT NULL,
-    created_date         timestamp    NOT NULL,
-    category             varchar(255) NOT NULL,
-    payment              int4         NOT NULL,
-    annual_interest_rate int4         NOT NULL,
-    payment_frequency    int4         NOT NULL,
-    compound_frequency   int4         NOT NULL,
-    years_of_growth      int4         NOT NULL,
-    start_amount         int8         NOT NULL,
+    id                   bigint           NOT NULL,
+    owner_id             bigint           NOT NULL,
+    basket_id            bigint           NOT NULL,
+    created_date         timestamp        NOT NULL,
+    category             varchar(255)     NOT NULL,
+    payment              integer          NOT NULL,
+    annual_interest_rate double precision NOT NULL,
+    payment_frequency    integer          NOT NULL,
+    compound_frequency   integer          NOT NULL,
+    years_of_growth      integer          NOT NULL,
+    start_amount         integer          NOT NULL,
     CONSTRAINT investments_pk PRIMARY KEY (id),
     CONSTRAINT investments_basket_fk FOREIGN KEY (basket_id) REFERENCES baskets (id) ON DELETE CASCADE,
     CONSTRAINT investments_users_fk FOREIGN KEY (owner_id) REFERENCES users (id)
